Separated files + fixed bug in checkmark
