+++ AtlasUI/AtlasUI/AddressPickerViewController.swift
@@ -12,7 +12,7 @@ enum AddressType {
 
 typealias AddressSelectionCompletion = (pickedAddress: Addressable, pickedAddressType: AddressType) -> Void
 
-final class AddressPickerViewController: UIViewController, UITableViewDelegate, UITableViewDataSource, CheckoutProviderType {
+final class AddressPickerViewController: UIViewController, CheckoutProviderType {
 
     internal var checkout: AtlasCheckout
     private let addressType: AddressType
@@ -20,13 +20,11 @@ final class AddressPickerViewController: UIViewController, UITableViewDelegate, 
 
     private let tableView = UITableView()
     private var addresses: [Address] = []
+    let tableviewDelegate: AddressListTableViewDelegate?
 
     var selectedAddress: Addressable? {
         didSet {
-            Async.main { [weak self] in
-                guard let strongSelf = self, selectedAddress = strongSelf.selectedAddress else { return }
-                strongSelf.selectionCompletion(pickedAddress: selectedAddress, pickedAddressType: strongSelf.addressType)
-            }
+            tableviewDelegate?.selectedAddress = selectedAddress
         }
     }
 
@@ -34,7 +32,10 @@ final class AddressPickerViewController: UIViewController, UITableViewDelegate, 
         addressSelectionCompletion: AddressSelectionCompletion) {
             self.checkout = checkout
             self.addressType = addressType
-            self.selectionCompletion = addressSelectionCompletion
+            selectionCompletion = addressSelectionCompletion
+            tableviewDelegate = AddressListTableViewDelegate(checkout: checkout,
+                addressType: addressType, addressSelectionCompletion: selectionCompletion)
+
             super.init(nibName: nil, bundle: nil)
     }
 
@@ -64,31 +65,11 @@ final class AddressPickerViewController: UIViewController, UITableViewDelegate, 
                     strongSelf.userMessage.show(error: error)
                 case .success(let addressList):
                     strongSelf.addresses = addressList.addresses
+                    strongSelf.tableviewDelegate?.addresses = addressList.addresses
                     strongSelf.tableView.reloadData()
                 }
             }
         }
-    }
-
-    func setupTableView() {
-
-        tableView.translatesAutoresizingMaskIntoConstraints = false
-        tableView.fillInSuperView()
-
-        tableView.delegate = self
-        tableView.dataSource = self
-        tableView.registerReusableCell(AddressRowViewCell.self)
-        tableView.rowHeight = UITableViewAutomaticDimension
-        tableView.estimatedRowHeight = 100
-        self.parentViewController?.navigationItem.rightBarButtonItem = self.editButtonItem()
-    }
-
-    func tableView(tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-        return addresses.count
-    }
-
-    func tableView(tableView: UITableView, canEditRowAtIndexPath indexPath: NSIndexPath) -> Bool {
-        return true
     }
 
     override func setEditing(editing: Bool, animated: Bool) {
@@ -96,41 +77,18 @@ final class AddressPickerViewController: UIViewController, UITableViewDelegate, 
         tableView.setEditing(editing, animated: animated)
     }
 
-    func tableView(tableView: UITableView, commitEditingStyle editingStyle: UITableViewCellEditingStyle,
-        forRowAtIndexPath indexPath: NSIndexPath) {
-            if editingStyle == .Delete {
-                let address = self.addresses[indexPath.item]
-                checkout.client.deleteAddress(address.id) { result in
-                    switch result {
-                    case .success(_):
-                        self.addresses.removeAtIndex(indexPath.item)
-                        tableView.deleteRowsAtIndexPaths([indexPath], withRowAnimation: .Fade)
-                    case .failure(let error):
-                        self.userMessage.show(error: error)
-                    }
-                }
-            }
-    }
+    func setupTableView() {
 
-    func tableView(tableView: UITableView, cellForRowAtIndexPath indexPath: NSIndexPath) -> UITableViewCell {
-        return tableView.dequeueReusableCell(AddressRowViewCell.self, forIndexPath: indexPath) { result in
-            switch result {
-            case let .dequeuedCell(addressRowCell):
-                let address = self.addresses[indexPath.item]
-                addressRowCell.address = address
-                addressRowCell.accessoryType = self.selectedAddress?.id == address.id ? .Checkmark : .None
-                return addressRowCell
-            case let .defaultCell(cell):
-                return cell
-            }
-        }
-    }
+        tableView.translatesAutoresizingMaskIntoConstraints = false
+        tableView.fillInSuperView()
 
-    func tableView(tableView: UITableView, didSelectRowAtIndexPath indexPath: NSIndexPath) {
-        selectedAddress = addresses[indexPath.item]
-
-        tableView.deselectRowAtIndexPath(indexPath, animated: true)
+        tableView.delegate = tableviewDelegate
+        tableView.dataSource = tableviewDelegate
+        tableView.registerReusableCell(AddressRowViewCell.self)
+        tableView.rowHeight = UITableViewAutomaticDimension
+        tableView.estimatedRowHeight = 100
         tableView.reloadData()
+        self.parentViewController?.navigationItem.rightBarButtonItem = self.editButtonItem()
     }
 
 }
